mock mapa + priprava na implementaci mock simulace

+++ diskretni_mapd_simulace/diskretni_mapd_simulace/Entities/Order.cs
@@ -12,7 +12,7 @@ namespace diskretni_mapd_simulace
         public Location currLocation;
         public Location targetLocation;
         public int state;
-        public int[] color;
+        public byte[] color;
 
 
         public enum states
